+++ CHANGELOG.md
@@ -9,3 +9,4 @@ None
 ## Changes
 
 - `user` command about current running bot restriction is removed
+- `mod general kick` and `mod general ban` will now filter out duplicates mentioned member